requirements.txt
Flask==2.0.3
python-dotenv==0.19.2
itsdangerous==2.0.1
Werkzeug==2.0.3
dbus-python==1.2.18
PyYAML==6.0
